pylog: Create printk subsystem to handle different ringbuffer formats
pylog command supports two formats of the internal kernel ringbuffer.
kernel-5.10 implements a new lockless ringbuffer and pylog will
need to support it.

Split the implementation of the current two formats as a preparation
step. It will better separate the completely different implementations.
